Move tests helper functions to dstack package for reusing
