+++ src/tests/_internal/server/background/tasks/test_process_pending_jobs.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from dstack._internal.core.models.runs import JobStatus
 from dstack._internal.server.background.tasks.process_pending_jobs import process_pending_jobs
-from tests._internal.server.common import (
+from dstack._internal.server.testing.common import (
     create_job,
     create_project,
     create_repo,
