+++ src/tests/_internal/server/routers/test_logs.py
@@ -9,7 +9,7 @@ from dstack._internal.core.models.users import GlobalRole, ProjectRole
 from dstack._internal.server import settings
 from dstack._internal.server.main import app
 from dstack._internal.server.services.projects import add_project_member
-from tests._internal.server.common import create_project, create_user, get_auth_headers
+from dstack._internal.server.testing.common import create_project, create_user, get_auth_headers
 
 client = TestClient(app)
