+++ src/tests/_internal/server/services/test_logs.py
@@ -8,7 +8,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from dstack._internal.server import settings
 from dstack._internal.server.schemas.runner import LogEvent
 from dstack._internal.server.services.logs import write_logs
-from tests._internal.server.common import create_project
+from dstack._internal.server.testing.common import create_project
 
 
 class TestWriteLogs:
